Refactor: Move logic to get default image uri to constants
Right now, to pull default image we pass empty string to
`image/PullBundle` function and have logic to get default image URI in
case the empty string for URI which is bit confusing. In this PR we are
adding `GetDefaultBundleImageRegistryURL` function to constants pkg to get
this info and change `image/PullBundle` signature.
